refactor existing prompts to allow for easier configuration
