+++ src/agent.ts
@@ -2,11 +2,11 @@ import {
   createOrchestratorRunner,
   OrchestratorRunner,
 } from './agents/workflows/orchestrator/orchestratorWorkflow.js';
-import { createPrompts } from './agents/workflows/orchestrator/prompts.js';
 import { createTools } from './agents/workflows/orchestrator/tools.js';
 import { createTwitterAgentTool } from './agents/workflows/twitter/twitterAgentTool.js';
 import { config } from './config/index.js';
 import { createTwitterApi } from './services/twitter/client.js';
+import { createPrompts } from './agents/workflows/orchestrator/prompts.js';
 
 const orchestatorConfig = async () => {
   const { USERNAME, PASSWORD, COOKIES_PATH } = config.twitterConfig;
@@ -14,8 +14,9 @@ const orchestatorConfig = async () => {
   const twitterAgent = createTwitterAgentTool(twitterApi);
 
   const { tools } = createTools();
-  const orchestratorPrompts = await createPrompts();
-  return { prompts: orchestratorPrompts, tools: [...tools, twitterAgent] };
+  const prompts = await createPrompts();
+
+  return { prompts, tools: [...tools, twitterAgent] };
 };
 
 const orchestratorConfig = await orchestatorConfig();
