+++ src/agents/workflows/orchestrator/nodes/messageSummaryNode.ts
@@ -24,7 +24,7 @@ export const createMessageSummaryNode = ({ orchestratorModel, prompts }: Orchest
         })
         .join('\n');
 
-      const formattedPrompt = await prompts.summaryPrompt.format({
+      const formattedPrompt = await prompts.messageSummaryPrompt.format({
         prevSummary,
         newMessages,
       });
