+++ src/agents/workflows/orchestrator/nodes/messageSummaryPrompt.ts
@@ -1,0 +1,44 @@
+import { ChatPromptTemplate, PromptTemplate } from '@langchain/core/prompts';
+import { SystemMessage } from '@langchain/core/messages';
+import { config } from '../../../../config/index.js';
+
+export const createMessageSummaryPrompt = async (customInstructions?: string) => {
+  const character = config.characterConfig;
+
+  const summarySystemPrompt = await PromptTemplate.fromTemplate(
+    `You are a helpful assistant that summarizes conversations in order to efficiently manage context window in AI model communication. You also have a personality.
+    {characterDescription}
+    {characterPersonality}
+
+    Instructions On OUTPUT:
+    - PRESERVE DETAILS
+    - SHOULD be detailed and contain all information conveyed in the messages
+    - SHOULD NOT omit any important details
+   
+    Custom Instructions:
+    {customInstructions}
+
+    THERE SHOULD BE LITTLE TO NO LOSS OF INFORMATION
+    
+    Format the summary in a clear, bulleted structure.`,
+  ).format({
+    characterDescription: character.description,
+    characterPersonality: character.personality,
+    customInstructions: customInstructions ?? 'None',
+  });
+
+  const messageSummaryPrompt = ChatPromptTemplate.fromMessages([
+    new SystemMessage(summarySystemPrompt),
+    [
+      'ai',
+      `Previous summary: {prevSummary}
+      
+      New AI messages to incorporate:
+      {newMessages}
+      
+      Create an updated DETAILED summary respected to INSTRUCTIONS ON OUTPUT.`,
+    ],
+  ]);
+
+  return messageSummaryPrompt;
+};
